Update AVI-Suite-NewUpdates.ipynb
Changes up until this point:
By "New Updates", we mean a way of updating in which all states in a "chain" of actions receive the same update. For example, if in state s action a1 is taken, then in state s+a1 action a2 is taken, and then in state s + a1 + a2 no action is taken, then the update rule performed for s+a1+a2 is also "fed back" and done for s and s+a1, replacing the previous approach which simply used updates of V(s+a1) for s, and V(s+a1+a2) for s+a1.

Additionally, update targets c + E{V(s')} - V(s0) or c + E{V(s')} - gt have been replaced by c + E{V(s')} - V(s0) - gt in the Cont-AVI algorithms; this seems to lead to better VFAs. This is done in both the old ContFull algorithm, and in our new algorithm. ContApprox has been deprecated, as it doesn't make sense to approximate E{V(s')} when we need to calculate it for the passive action anyway. However, APE-FA-ContApprox remains in use, as this doesn't require the calculation of expected values to find actions.

We also introduce gEvaluationContFixed, which takes a parameter g0 which replaces g in the action selection, so that the same policy is used throughout.

Some investigation in Exact DP reveals that, in some situations, the value function can in fact become negative. Such a situation can be constructed as follows: consider the 5 link inhomogeneous problem. Create a policy which is fully passive when link 3 is healthy, and is close to optimal (or at least reasonable) when 3 is damaged, and never repairs 3. Starting in state s0 = (1,1,1,1,1) then means we have to "suffer through" the use of a bad policy until link 3 becomes broken, past which we never repair 3 and use a much better policy. State (1,1,3,1,1) therefore becomes better than s0, and therefore recieves a negative value function when s0 is the reference state.

Due to this investigation, the rule in the AVI/APE functions which ensured update targets vHat became max(vHat,0) has been removed.