Update Problem-Suite-Experiments.ipynb
Found big dumb mistake where, when updating across the state trace, the substates passed to the updateVFA function were always that of the current state. This can been fixed, and all tests need to be re-run.

smarviNewVFA now returns current state and stateTrace, so it can more easily be resumed.